Move DashboardClickDrill logic to metabase-lib (#25892)
